Resource-based authorization for TimesheetEntries

+++ Timesheets/Controllers/TimesheetEntriesController.cs
@@ -1,33 +1,122 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Timesheets.Areas.Identity.Data;
 using Timesheets.Data;
 using Timesheets.Models;
 
 namespace Timesheets.Controllers
 {
+    // https://docs.microsoft.com/en-us/aspnet/core/security/authorization/resourcebased?view=aspnetcore-2.1
+    // https://docs.microsoft.com/en-us/aspnet/core/security/authorization/policies?view=aspnetcore-2.1#applying-policies-to-razor-pages
     public class TimesheetEntriesController : Controller
     {
+        private readonly IAuthorizationService _authorizationService;
         private readonly ApplicationDbContext _context;
 
-        public TimesheetEntriesController(ApplicationDbContext context)
+        public TimesheetEntriesController(IAuthorizationService authorizationService, ApplicationDbContext context)
         {
+            _authorizationService = authorizationService;
             _context = context;
         }
 
         // GET: TimesheetEntries
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.TimesheetEntries.Include(t => t.Project).Include(t => t.User);
-            return View(await applicationDbContext.ToListAsync());
+            if (! User.Identity.IsAuthenticated)
+            {
+                return new ChallengeResult();
+            }
+
+            var timesheetList = _context.TimesheetEntries.
+                Where(t => t.UserId == User.FindFirst(ClaimTypes.NameIdentifier).Value).
+                Include(t => t.Project).Include(t => t.User);
+            return View(await timesheetList.ToListAsync());
         }
 
         // GET: TimesheetEntries/Details/5
         public async Task<IActionResult> Details(long? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var timesheetEntry = await _context.TimesheetEntries
+                .Include(t => t.Project)
+                .Include(t => t.User)
+                .FirstOrDefaultAsync(m => m.TimesheetEntryId == id);
+
+            if (timesheetEntry == null)
+            {
+                return NotFound();
+            }
+
+            var authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, timesheetEntry, "SameTimesheetEntryCreator");
+            if (authorizationResult.Succeeded)
+            {
+                return View(timesheetEntry);
+            }
+            else if (User.Identity.IsAuthenticated)
+            {
+                return new ForbidResult();
+            }
+            else
+            {
+                return new ChallengeResult();
+            }
+        }
+
+        // GET: TimesheetEntries/Create
+        public IActionResult Create()
+        {
+            //ViewData["ProjectId"] = new SelectList(_context.Projects, "ProjectId", "ProjectId");
+            //ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id");
+
+            ViewBag.Projects = new SelectList(_context.Projects, "ProjectId", "Name");
+            ViewBag.ApplicationUsers = new SelectList(_context.ApplicationUsers
+               .Select(u => new { FullName = String.Format("{0} {1}", u.FirstName, u.LastName), u.Id })
+               , "Id", "FullName");
+            // TODO: automatically use current user's id
+            return View();
+        }
+
+        // POST: TimesheetEntries/Create
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Create( TimesheetEntry timesheetEntry)
+        {
+            var justCheck = _context.TimesheetEntries
+                .Where(t => t.DateCreated.Date == timesheetEntry.DateCreated.Date && t.ProjectId == timesheetEntry.ProjectId).FirstOrDefault();
+
+            if (ModelState.IsValid && justCheck == null && timesheetEntry.HoursWorked>0)
+            {
+                _context.Add(timesheetEntry);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            //ViewData["ProjectId"] = new SelectList(_context.Projects, "ProjectId", "ProjectId", timesheetEntry.ProjectId);
+            //ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id", timesheetEntry.UserId);
+            ViewBag.Projects = new SelectList(_context.Projects, "ProjectId", "Name");
+            ViewBag.ApplicationUsers = new SelectList(_context.ApplicationUsers
+               .Select(u => new { FullName = String.Format("{0} {1}", u.FirstName, u.LastName), u.Id })
+               , "Id", "FullName");
+
+            return View(timesheetEntry);
+        }
+
+        // GET: TimesheetEntries/Edit/5
+        public async Task<IActionResult> Edit(long? id)
         {
             if (id == null)
             {
@@ -43,68 +132,24 @@ namespace Timesheets.Controllers
                 return NotFound();
             }
 
-            return View(timesheetEntry);
-        }
-
-        // GET: TimesheetEntries/Create
-        public IActionResult Create()
-        {
-            //ViewData["ProjectId"] = new SelectList(_context.Projects, "ProjectId", "ProjectId");
-            //ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id");
-
-            ViewBag.Projects = new SelectList(_context.Projects, "ProjectId", "Name");
-            ViewBag.ApplicationUsers = new SelectList(_context.ApplicationUsers
-               .Select(u => new { FullName = String.Format("{0} {1}", u.FirstName, u.LastName), u.Id })
-               , "Id", "FullName");
-            return View();
-        }
-
-        // POST: TimesheetEntries/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create( TimesheetEntry timesheetEntry)
-        {
-            var justCheck = _context.TimesheetEntries
-                .Where(t => t.DateCreated.Date == timesheetEntry.DateCreated.Date && t.ProjectId == timesheetEntry.ProjectId).FirstOrDefault();
-
-            //var applicationDbContext = _context.TimesheetEntries.Include(t => t.Project).Include(t => t.User);
-            //    return View(await applicationDbContext.ToListAsync());
-            if (ModelState.IsValid && justCheck == null && timesheetEntry.HoursWorked>0)
+            var authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, timesheetEntry, "SameTimesheetEntryCreator");
+            if (authorizationResult.Succeeded)
             {
-                _context.Add(timesheetEntry);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                ViewBag.Projects = new SelectList(_context.Projects, "ProjectId", "Name");
+                ViewBag.ApplicationUsers = new SelectList(_context.ApplicationUsers
+                   .Select(u => new { FullName = String.Format("{0} {1}", u.FirstName, u.LastName), u.Id })
+                   , "Id", "FullName");
+                return View(timesheetEntry);
             }
-            //ViewData["ProjectId"] = new SelectList(_context.Projects, "ProjectId", "ProjectId", timesheetEntry.ProjectId);
-            //ViewData["UserId"] = new SelectList(_context.ApplicationUsers, "Id", "Id", timesheetEntry.UserId);
-            ViewBag.Projects = new SelectList(_context.Projects, "ProjectId", "Name");
-            ViewBag.ApplicationUsers = new SelectList(_context.ApplicationUsers
-               .Select(u => new { FullName = String.Format("{0} {1}", u.FirstName, u.LastName), u.Id })
-               , "Id", "FullName");
-            return View(timesheetEntry);
-        }
-
-        // GET: TimesheetEntries/Edit/5
-        public async Task<IActionResult> Edit(long? id)
-        {
-            if (id == null)
+            else if (User.Identity.IsAuthenticated)
             {
-                return NotFound();
+                return new ForbidResult();
             }
-
-            var timesheetEntry = await _context.TimesheetEntries.FindAsync(id);
-            if (timesheetEntry == null)
+            else
             {
-                return NotFound();
+                return new ChallengeResult();
             }
-
-            ViewBag.Projects = new SelectList(_context.Projects, "ProjectId", "Name");
-            ViewBag.ApplicationUsers = new SelectList(_context.ApplicationUsers
-               .Select(u => new { FullName = String.Format("{0} {1}", u.FirstName, u.LastName), u.Id })
-               , "Id", "FullName");
-            return View(timesheetEntry);
         }
 
         // POST: TimesheetEntries/Edit/5
@@ -118,8 +163,9 @@ namespace Timesheets.Controllers
             {
                 return NotFound();
             }
-            
-            if (timesheetEntry.HoursWorked > 0 )
+            var authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, timesheetEntry, "SameTimesheetEntryCreator");
+            if (authorizationResult.Succeeded && timesheetEntry.HoursWorked > 0 )
             {
                 try
                 {
@@ -162,8 +208,20 @@ namespace Timesheets.Controllers
             {
                 return NotFound();
             }
-
-            return View(timesheetEntry);
+            var authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, timesheetEntry, "SameTimesheetEntryCreator");
+            if (authorizationResult.Succeeded)
+            {
+                return View(timesheetEntry);
+            }
+            else if (User.Identity.IsAuthenticated)
+            {
+                return new ForbidResult();
+            }
+            else
+            {
+                return new ChallengeResult();
+            }
         }
 
         // POST: TimesheetEntries/Delete/5
@@ -171,10 +229,20 @@ namespace Timesheets.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(long id)
         {
+            if (id == null)
+            {
+                return NotFound();
+            }
             var timesheetEntry = await _context.TimesheetEntries.FindAsync(id);
-            _context.TimesheetEntries.Remove(timesheetEntry);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var authorizationResult = await _authorizationService
+                   .AuthorizeAsync(User, timesheetEntry, "SameTimesheetEntryCreator");
+            if (authorizationResult.Succeeded)
+            {
+                _context.TimesheetEntries.Remove(timesheetEntry);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return View();
         }
 
         private bool TimesheetEntryExists(long id)
